fast/mediastream/getUserMedia-to-canvas-1.html and fast/mediastream/getUserMedia-to-canvas-2.html are failing on MacOS when remote video frame flag is enabled https://bugs.webkit.org/show_bug.cgi?id=237042 <rdar://problem/89298601>
Reviewed by Eric Carlson.

Video element videoWidth/videoHeight are updated when track settings are changed,
and not when video frames being received.

This does not guarantee that videoWidth/videoHeight (main thread) are fully in sync
with the actual video frames (background thread) when there is a change of size.

We might be able to tighten this in MediaPlayerPrivateMediaStreamAVFObjC by only resorting on video frames.
In the meantime, we update the test to use requestVideoFrameCallback which provides accurate per frame metadata.

* fast/mediastream/getUserMedia-to-canvas-1.html:
* fast/mediastream/getUserMedia-to-canvas-2.html:
* webrtc/routines.js:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290819 268f45cc-cd09-0410-ab3c-d52691b4dbfc