[iOS] Unable to scroll to a found text range when there is an existing selection https://bugs.webkit.org/show_bug.cgi?id=237407 rdar://89653213
Reviewed by Wenson Hsieh.

Source/WebCore:

Scrolling to a found text range is performed by creating a
`TemporarySelectionChange` and using the `RevealSelectionBounds`
`TemporarySelectionOption`. When a `TemporarySelectionChange` is
destroyed, the original selection is restored.

Currently, the selection is restored using the same set of selection
options used to make the temporary selection. Consequently, whenever a
"reveal" option is specified, WebKit scrolls to reveal the temporary
selection, and then scrolls again to reveal the original selection.
This behavior means that an attempt to scroll to a found text range
will fail if the document has an existing selection.

To fix, do not add any of the "reveal" options to the set of selection
options when restoring the original selection during a
`TemporarySelectionChange`. The only other features that use
`TemporarySelectionChange` with "reveal" options are App Highlights
and Scroll To Text Fragment. Neither of these features require WebKit
to scroll to the original selection.

* editing/Editor.cpp:
(WebCore::TemporarySelectionChange::TemporarySelectionChange):
(WebCore::TemporarySelectionChange::~TemporarySelectionChange):
(WebCore::TemporarySelectionChange::setSelection):
* editing/Editor.h:

Tools:

Add an API test that sets a selection in the document, finds some
text outside the viewport, and scrolls to make the found text
visible.

* TestWebKitAPI/Tests/WebKitCocoa/FindInPage.mm:
(-[TestScrollViewDelegate init]):
(-[TestScrollViewDelegate scrollViewDidEndScrollingAnimation:]):
(TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290834 268f45cc-cd09-0410-ab3c-d52691b4dbfc