"OffscreenCanvas" in IDLs doesn't seem to be able to be compiled https://bugs.webkit.org/show_bug.cgi?id=232733 <rdar://problem/85318653>
Patch by Dan Glastonbury <djg@apple.com> on 2022-03-04
Reviewed by Sam Weinig.

OffscreenCanvas is controlled by ENABLE flags. Add conditional
compilation to handle OffscreenCanvas to the WebGPU IDL bindings
when the OffscreenCanvas feature is enabled.

* Modules/WebGPU/GPUCanvasContext.cpp:
(WebCore::GPUCanvasContext::canvas):
* Modules/WebGPU/GPUCanvasContext.h:
* Modules/WebGPU/GPUCanvasContext.idl:
Add OffscreenCanvas to canvas attribute when it is enabled.
* Modules/WebGPU/GPUImageCopyExternalImage.h:
* Modules/WebGPU/GPUImageCopyExternalImage.idl:
Add OffscreenCanvas to source attribute when it is enabled.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290838 268f45cc-cd09-0410-ab3c-d52691b4dbfc