[GPU Process] Canvas compositing buffer should be created through its GraphicsContext https://bugs.webkit.org/show_bug.cgi?id=237260 rdar://89196918
Reviewed by Simon Fraser.

Source/WebCore:

If the backend of the underlying ImageBuffer of the canvas is remote the
compositing ImageBuffer will also be remote. This will transfer the whole
compositing operation to GPUProcess.

The layout test fast/canvas/canvas-composite-canvas.html crashes because
of this bug on the GPUP layout bots.

* html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::CanvasRenderingContext2DBase::fullCanvasCompositedDrawImage):
(WebCore::CanvasRenderingContext2DBase::createCompositingBuffer): Deleted.
* html/canvas/CanvasRenderingContext2DBase.h:

LayoutTests:

* gpu-process/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290839 268f45cc-cd09-0410-ab3c-d52691b4dbfc