[model] <model interactive> doesn't allow model to be rotated on iOS https://bugs.webkit.org/show_bug.cgi?id=237831 rdar://89698247
Reviewed by Simon Fraser.

Source/WebCore:

A <model> element is *not* interactive by default and the "interactive" HTML attribute
must be set explicitly. We used to mirror this initial disabled state by explicitly
setting "userInteractionEnabled" to NO on WKModelView during its creation, and we would
update that property when HTMLModelElement::isInteractive() would change value.

However, in the case where the attribute was originally set, we would completely
disregard that value.

We now account for it when we create the GraphicsLayer for the <model> element, letting the
existing GraphicsLayer machinery to set the matching "userInteractionEnabled" property on
the WKModelView.

* Modules/model-element/HTMLModelElement.h:
* platform/graphics/GraphicsLayer.h:
(WebCore::GraphicsLayer::setContentsToModel):
* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::setContentsToModel):
* platform/graphics/ca/GraphicsLayerCA.h:
* rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::updateConfiguration):

Source/WebKit:

Remove the call to set userInteractionEnabled on the WKModelView when created. This property
is now set via the layer tree application code from the GraphicsLayer::userInteractionEnabled()
value.

* UIProcess/ios/WKModelView.mm:
(-[WKModelView initWithModel:]):

LayoutTests:

Add a test where the <model> element is added to the DOM with the "interactive"
HTML attribute already set.

* model-element/model-element-interactive-dragging-expected.txt:
* model-element/model-element-interactive-dragging.html:
* model-element/resources/model-utils.js:
(const.makeModel):
(const.readyModel.async test):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291245 268f45cc-cd09-0410-ab3c-d52691b4dbfc